# Conflicts:
#	epc-tendering-service/src/main/java/com/epc/tendering/service/service/project/impl/TProjectBasicInfoServiceImpl.java

+++ epc-platform-service/src/main/java/com/epc/platform/service/service/biddingagency/impl/AgencyServiceImpl.java
@@ -1,7 +1,6 @@
 package com.epc.platform.service.service.biddingagency.impl;
 import com.epc.administration.facade.biddingagency.handle.*;
 import com.epc.administration.facade.biddingagency.vo.AgencyAttachmentVO;
-
 import com.epc.administration.facade.biddingagency.dto.QueryDetailIfo;
 import com.epc.administration.facade.biddingagency.vo.AgencyUserAttachmentVO;
 import com.epc.administration.facade.biddingagency.vo.BiddingAgencyVO;
@@ -64,11 +63,9 @@ public class AgencyServiceImpl implements AgencyService {
         try {
             return Result.success(tAgencyBasicInfoMapper.insertSelective(tAgencyBasicInfo) > 0);
         } catch (BusinessException e) {
-            LOGGER.error("BusinessException inserttAgencyBasicInfo : {}", e);
+            LOGGER.error("BusinessException insertBiddingAgencyBasicInfo : {}", e);
+            TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
             return Result.error(ErrorMessagesEnum.INSERT_FAILURE);
-        } catch (Exception e) {
-            LOGGER.error("BusinessException inserttAgencyBasicInfo : {}", e);
-            return Result.error(e.getMessage());
         }
     }
     /**
@@ -79,6 +76,7 @@ public class AgencyServiceImpl implements AgencyService {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public Result<Boolean> insertBiddingAgencyDetailInfo(BiddingHandle biddingHandle) {
+
         TAgencyDetailInfo detailInfo = new TAgencyDetailInfo();
         Date date = new Date();
         detailInfo.setAgencyId(biddingHandle.getId());
@@ -123,11 +121,11 @@ public class AgencyServiceImpl implements AgencyService {
             tAgencyBasicInfo.setState(Const.STATE.COMMITTED);
             return Result.success(tAgencyBasicInfoMapper.updateByPrimaryKeySelective(tAgencyBasicInfo)>0);
         }catch (BusinessException e) {
-            LOGGER.error("BusinessException inserttAgencyDetailInfo : {}", e);
+            LOGGER.error("BusinessException insertBiddingAgencyDetailInfo : {}", e);
             TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
             return Result.error(ErrorMessagesEnum.INSERT_FAILURE);
         }catch (Exception e){
-            LOGGER.error("BusinessException inserttAgencyDetailInfo : {}", e);
+            LOGGER.error("BusinessException insertBiddingAgencyDetailInfo : {}", e);
             TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
             return Result.error(e.getMessage());
         }
@@ -144,6 +142,7 @@ public class AgencyServiceImpl implements AgencyService {
         TAgencyDetailInfoCriteria tAgencyDetailInfoCriteria = new TAgencyDetailInfoCriteria();
         tAgencyDetailInfoCriteria.createCriteria().andAgencyIdEqualTo(biddingHandle.getId());
         Date date = new Date();
+        //详情
         detailInfo.setAgencyId(biddingHandle.getId());
         detailInfo.setCompanyName(biddingHandle.getCompanyName());
         detailInfo.setUniformCreditCode(biddingHandle.getUniformCreditCode());
@@ -153,11 +152,18 @@ public class AgencyServiceImpl implements AgencyService {
         TAgencyAttachment attachment = new TAgencyAttachment();
         attachment.setUpdateAt(date);
         try {
+            List<TAgencyDetailInfo> tAgencyDetailInfos = tAgencyDetailInfoMapper.selectByExample(tAgencyDetailInfoCriteria);
+            if(tAgencyDetailInfos.isEmpty()){
+                tAgencyDetailInfoMapper.insertSelective(detailInfo);
+            }else {
+                tAgencyDetailInfoMapper.updateByExampleSelective(detailInfo,tAgencyDetailInfoCriteria);
+            }
             TAgencyAttachmentCriteria criteria = new TAgencyAttachmentCriteria();
             criteria.createCriteria().andAgencyIdEqualTo(biddingHandle.getId());
-            tAgencyDetailInfoMapper.deleteByExample(tAgencyDetailInfoCriteria);
-            tAgencyAttachmentMapper.deleteByExample(criteria);
-            tAgencyDetailInfoMapper.insertSelective(detailInfo);
+            List<TAgencyAttachment> tAgencyAttachments = tAgencyAttachmentMapper.selectByExample(criteria);
+            if(!tAgencyAttachments.isEmpty()){
+                tAgencyAttachmentMapper.deleteByExample(criteria);
+            }
             //法人身份证反面照片url
             attachment.setCertificateType(AttachmentEnum.LEGAL_ID_CARD_OTHER.getCode());
             attachment.setCertificateFilePath(biddingHandle.getLegalIdCardOther());
@@ -180,16 +186,17 @@ public class AgencyServiceImpl implements AgencyService {
                     tAgencyAttachmentMapper.insertSelective(attachment);
                 }
             }
+            //基础信息更新
             TAgencyBasicInfo tAgencyBasicInfo = new TAgencyBasicInfo();
             tAgencyBasicInfo.setId(biddingHandle.getId());
             tAgencyBasicInfo.setState(Const.STATE.COMMITTED);
             return Result.success(tAgencyBasicInfoMapper.updateByPrimaryKeySelective(tAgencyBasicInfo)>0);
         }catch (BusinessException e) {
-            LOGGER.error("BusinessException updateByPrimaryKeySelective : {}", e);
+            LOGGER.error("BusinessException updateBiddingAgencyDetailInfo : {}", e);
             TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
             return Result.error(ErrorMessagesEnum.UPDATE_FAILURE);
         }catch (Exception e){
-            LOGGER.error("BusinessException updateByPrimaryKeySelective : {}", e);
+            LOGGER.error("BusinessException updateBiddingAgencyDetailInfo : {}", e);
             TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
             return Result.error(ErrorMessagesEnum.UPDATE_FAILURE);
         }
@@ -207,7 +214,7 @@ public class AgencyServiceImpl implements AgencyService {
         try{
             return Result.success(tAgencyBasicInfoMapper.updateByPrimaryKeySelective(tAgencyBasicInfo)>0);
         }catch (BusinessException e){
-            LOGGER.error("BusinessException updateByPrimaryKeySelective : {}", e);
+            LOGGER.error("BusinessException deleteBiddingAgencyDetailInfo : {}", e);
             return Result.error(ErrorMessagesEnum.UPDATE_FAILURE);
         }
     }
@@ -218,23 +225,24 @@ public class AgencyServiceImpl implements AgencyService {
      */
     @Override
     public Result<AgencyUserAttachmentVO> queryBiddingAgencyDetailInfo(Long whereId) {
-
             try {
+                //查询对应基础信息
                 TAgencyBasicInfo tAgencyBasicInfo = tAgencyBasicInfoMapper.selectByPrimaryKey(whereId);
-
                 TAgencyDetailInfoCriteria criteria = new TAgencyDetailInfoCriteria();
                 criteria.createCriteria().andAgencyIdEqualTo(whereId);
-
-
-
-                TAgencyAttachmentCriteria tAgencyAttachmentCriteria = new TAgencyAttachmentCriteria();
-                tAgencyAttachmentCriteria.createCriteria().andAgencyIdEqualTo(whereId);
-
-                List<TAgencyAttachment> tAgencyAttachments = tAgencyAttachmentMapper.selectByExample(tAgencyAttachmentCriteria);
-
+                List<TAgencyDetailInfo> tAgencyDetailInfos = tAgencyDetailInfoMapper.selectByExample(criteria);
+                //如果未完善信息 返回基础信息
+                if(tAgencyDetailInfos.isEmpty()){
+                    AgencyUserAttachmentVO agencyUserAttachmentVO = new AgencyUserAttachmentVO();
+                    agencyUserAttachmentVO.setName(tAgencyBasicInfo.getName());
+                    agencyUserAttachmentVO.setCellphone(tAgencyBasicInfo.getCellphone());
+                    agencyUserAttachmentVO.setState(tAgencyBasicInfo.getState());
+                    return Result.success(agencyUserAttachmentVO);
+                }
+                //装填用户所有信息开始
+                TAgencyDetailInfo tAgencyDetailInfo = tAgencyDetailInfos.get(0);
                 AgencyUserAttachmentVO agencyUserAttachmentVO = new AgencyUserAttachmentVO();
                 agencyUserAttachmentVO.setId(whereId);
-                TAgencyDetailInfo tAgencyDetailInfo = tAgencyDetailInfoMapper.selectByExample(criteria).get(0);
                 if(tAgencyDetailInfo!=null){
                     agencyUserAttachmentVO.setCompanyName(tAgencyDetailInfo.getCompanyName());
                     agencyUserAttachmentVO.setUniformCreditCode(tAgencyDetailInfo.getUniformCreditCode());
@@ -242,6 +250,9 @@ public class AgencyServiceImpl implements AgencyService {
                     agencyUserAttachmentVO.setPublicBanAccountNumber(tAgencyDetailInfo.getPublicBanAccountNumber());
 
                 }
+                TAgencyAttachmentCriteria tAgencyAttachmentCriteria = new TAgencyAttachmentCriteria();
+                tAgencyAttachmentCriteria.createCriteria().andAgencyIdEqualTo(whereId);
+                List<TAgencyAttachment> tAgencyAttachments = tAgencyAttachmentMapper.selectByExample(tAgencyAttachmentCriteria);
                 agencyUserAttachmentVO.setCreateAt(new Date());
                 agencyUserAttachmentVO.setIsDeleted(tAgencyBasicInfo.getIsDeleted());
                 agencyUserAttachmentVO.setCellphone(tAgencyBasicInfo.getCellphone());
@@ -250,16 +261,23 @@ public class AgencyServiceImpl implements AgencyService {
                 List<AgencyAttachmentVO> agencyAttachmentVOS = new ArrayList<>();
 
                 for (TAgencyAttachment tAgencyAttachment : tAgencyAttachments) {
+                    if(tAgencyAttachment.getCertificateType().equals(AttachmentEnum.LEGAL_ID_CARD_OTHER.getCode())){
+                        agencyUserAttachmentVO.setLegalIdCardOther(tAgencyAttachment.getCertificateFilePath());
+                    }  else if (tAgencyAttachment.getCertificateType().equals(AttachmentEnum.LEGAL_ID_CARD_POSITIVE.getCode())){
+                        agencyUserAttachmentVO.setLegalIdCardPositive(tAgencyAttachment.getCertificateFilePath());
+                    }   else if (tAgencyAttachment.getCertificateType().equals(AttachmentEnum.BUSINESS_LICENSE.getCode())){
+                        agencyUserAttachmentVO.setBusinessLicense(tAgencyAttachment.getCertificateFilePath());
+                    }
                     AgencyAttachmentVO agencyAttachmentVO = new AgencyAttachmentVO();
                     agencyAttachmentVO.setCertificateFilePath(tAgencyAttachment.getCertificateFilePath());
                     agencyAttachmentVO.setCertificateName(tAgencyAttachment.getCertificateName());
-                    agencyAttachmentVO.setCertificateType(tAgencyAttachment.getCertificateType());
                     agencyAttachmentVOS.add(agencyAttachmentVO);
                 }
                 agencyUserAttachmentVO.setAgencyAttachmentVOS(agencyAttachmentVOS);
+                //装填用户所有信息结束
                 return Result.success(agencyUserAttachmentVO);
             } catch (BusinessException e) {
-                LOGGER.error("BusinessException deleteByPrimaryKey : {}", e);
+                LOGGER.error("BusinessException queryBiddingAgencyDetailInfo : {}", e);
                 return Result.error(ErrorMessagesEnum.SELECT_FAILURE);
             }
     }
